[tests] Pull out random number generator
