+++ tests/common/mod.rs
@@ -12,6 +12,7 @@ pub mod input_arg;
 pub mod output_option;
 pub mod process;
 pub mod program;
+pub mod random;
 pub mod target;
 pub mod test_dir;
 pub mod thin;
